refactor(master &  authnode): split package rocksdb from raftstore to avoid import by other.
