+++ raftstore/raftstore_db/store_rocksdb.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package raftstore
+package raftstore_db
 
 import (
 	"fmt"
